Extracted ErrorMessage into relevant class files, Updated imports, added formatting
