+++ lib/msal-core/src/Authority.ts
@@ -1,7 +1,7 @@
 import { IUri } from "./IUri";
 import { Utils } from "./Utils";
 import { ITenantDiscoveryResponse } from "./ITenantDiscoveryResponse";
-import { ErrorMessage } from "./error/ErrorMessage";
+import { ClientConfigurationErrorMessage } from "./error/ClientConfigurationError";
 import { XhrClient } from "./XHRClient";
 
 /**
@@ -117,15 +117,15 @@ export abstract class Authority {
     try {
       components = this.CanonicalAuthorityUrlComponents;
     } catch (e) {
-      throw ErrorMessage.invalidAuthorityType;
+      throw ClientConfigurationErrorMessage.invalidAuthorityType;
     }
 
     if (!components.Protocol || components.Protocol.toLowerCase() !== "https:") {
-      throw ErrorMessage.authorityUriInsecure;
+      throw ClientConfigurationErrorMessage.authorityUriInsecure;
     }
 
     if (!components.PathSegments || components.PathSegments.length < 1) {
-      throw ErrorMessage.authorityUriInvalidPath;
+      throw ClientConfigurationErrorMessage.authorityUriInvalidPath;
     }
   }
