+++ lib/msal-core/src/B2cAuthority.ts
@@ -23,7 +23,7 @@
 
 import { AadAuthority } from "./AadAuthority";
 import { Authority, AuthorityType } from "./Authority";
-import { ErrorMessage } from "./error/ErrorMessage";
+import { ClientConfigurationErrorMessage } from "./error/ClientConfigurationError";
 import { Utils } from "./Utils";
 
 /**
@@ -36,7 +36,7 @@ export class B2cAuthority extends AadAuthority {
 
     let pathSegments = urlComponents.PathSegments;
     if (pathSegments.length < 3) {
-        throw ErrorMessage.b2cAuthorityUriInvalidPath;
+        throw ClientConfigurationErrorMessage.b2cAuthorityUriInvalidPath;
     }
 
     this.CanonicalAuthority = `https://${urlComponents.HostNameAndPort}/${pathSegments[0]}/${pathSegments[1]}/${pathSegments[2]}/`;
@@ -62,6 +62,6 @@ export class B2cAuthority extends AadAuthority {
     }
 
     return new Promise<string>((resolve, reject) =>
-      reject(ErrorMessage.unsupportedAuthorityValidation));
+      reject(ClientConfigurationErrorMessage.unsupportedAuthorityValidation));
   }
 }
