+++ lib/msal-core/src/index.ts
@@ -3,7 +3,6 @@ export { Logger } from "./Logger";
 export { LogLevel } from "./Logger";
 export { User } from "./User";
 export { Constants } from "./Constants";
-export { ErrorMessage } from "./error/ErrorMessage";
 export { TokenResponse } from "./RequestInfo";
 export { Authority } from "./Authority";
 export { CacheResult } from "./UserAgentApplication";
@@ -13,5 +12,5 @@ export { AuthError } from "./error/AuthError";
 export { ClientAuthError } from "./error/ClientAuthError";
 export { ServerError } from "./error/ServerError";
 export { ClientConfigurationError } from "./error/ClientConfigurationError";
-export { InteractionRequiredAuthError } from "./error/InteractionRequiredError";
+export { InteractionRequiredAuthError } from "./error/InteractionRequiredAuthError";
 export { ClaimsRequiredAuthError } from "./error/ClaimsRequiredError";
